organized models into cheese and user packages and refactored import statements accordingly
