+++ src/main/java/org/launchcode/cheesemvc/controllers/CheeseController.java
@@ -1,8 +1,8 @@
 package org.launchcode.cheesemvc.controllers;
 
-import org.launchcode.cheesemvc.models.Cheese;
-import org.launchcode.cheesemvc.models.CheeseData;
-import org.launchcode.cheesemvc.models.CheeseType;
+import org.launchcode.cheesemvc.models.cheese.Cheese;
+import org.launchcode.cheesemvc.models.cheese.CheeseData;
+import org.launchcode.cheesemvc.models.cheese.CheeseType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
