+++ src/main/java/org/launchcode/cheesemvc/models/cheese/Cheese.java
@@ -1,4 +1,4 @@
-package org.launchcode.cheesemvc.models;
+package org.launchcode.cheesemvc.models.cheese;
 
 import javax.validation.constraints.*;
 import java.util.Objects;
