+++ src/main/java/org/launchcode/cheesemvc/models/cheese/CheeseData.java
@@ -1,4 +1,4 @@
-package org.launchcode.cheesemvc.models;
+package org.launchcode.cheesemvc.models.cheese;
 
 import java.util.ArrayList;
