+++ src/main/java/org/launchcode/cheesemvc/models/cheese/CheeseType.java
@@ -1,4 +1,4 @@
-package org.launchcode.cheesemvc.models;
+package org.launchcode.cheesemvc.models.cheese;
 
 public enum CheeseType {
     HARD ("Hard"),
